extract governance team options to a separate component
